removing warning for 64 bit compile.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/utils@77 720011fe-810a-0410-ac5b-998ee3860a72

+++ geom/CD-libs/PQP_v1.2/src/PQP.cpp
@@ -251,10 +251,10 @@ PQP_Model::MemUsage(int msg)
   if (msg) 
   {
     fprintf(stderr,"Total for model %p: %d bytes\n", this, total_mem);
-    fprintf(stderr,"BVs: %d alloced, take %d bytes each\n", 
-            num_bvs, sizeof(BV));
-    fprintf(stderr,"Tris: %d alloced, take %d bytes each\n", 
-            num_tris, sizeof(Tri));
+    fprintf(stderr,"BVs: %d alloced, take %u bytes each\n", 
+            num_bvs, (unsigned int)sizeof(BV));
+    fprintf(stderr,"Tris: %d alloced, take %u bytes each\n", 
+            num_tris, (unsigned int)sizeof(Tri));
   }
   
   return total_mem;
